refactor(pass): declare the query run's passes - declare the run each pass serves as a fact next to indexing and awake, so a query resolve opens the passes declared for it instead of a hand written replay of the step - move the query kernel out of commit into its own queries module and declare query_aabbs and query as passes, driving both runs through one schedule with one pass order, one validation and one observable run - scope host edit consumption: a query resolve previews the idempotent scene edits and leaves the one shot force, torque and impulse inputs pending, so only a step consumes them - fix a force or torque applied before a query resolve being wiped by the next step's input clear, and cover pending inputs, a sleeping body, patch replay and query pass coverage

+++ crates/dynamis-broadphase/src/domain.rs
@@ -58,7 +58,7 @@ impl Domain for BroadphaseDomain {
         Broadphase::new(context, streams, passes)
     }
 
-    fn gates(_: &()) -> u8 {
+    fn gates(_: &()) -> u16 {
         0
     }
